+++ src/main/java/roomescape/domain/home/PageController.java
@@ -1,4 +1,4 @@
-package roomescape.user.home;
+package roomescape.domain.home;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
