+++ src/main/java/roomescape/domain/reservation/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.user.reservation;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -9,8 +9,8 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.user.reservationtime.ReservationTime;
-import roomescape.user.theme.Theme;
+import roomescape.domain.reservationtime.ReservationTime;
+import roomescape.domain.theme.Theme;
 
 @Repository
 public class ReservationRepository {
