+++ src/main/java/roomescape/domain/reservationtime/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.user.reservationtime;
+package roomescape.domain.reservationtime;
 
 import java.time.LocalTime;
