+++ src/main/java/roomescape/domain/reservationtime/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.user.reservationtime;
+package roomescape.domain.reservationtime;
 
 
 import java.util.List;
