+++ src/main/java/roomescape/domain/theme/ThemeController.java
@@ -1,11 +1,11 @@
-package roomescape.user.theme;
+package roomescape.domain.theme;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.user.theme.dto.ThemeResponse;
+import roomescape.domain.theme.dto.ThemeResponse;
 
 @RestController
 @RequestMapping("/themes")
